requirements.txt
Django==4.2.23
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
python-dotenv==1.0.0
ollama==0.1.6
PyPDF2==3.0.1
python-docx==1.1.0
Pillow==10.2.0
celery==5.3.4
redis==5.0.1
requests==2.31.0



